requirements.txt
pillow
webcolors
requests
Beautifulsoup4
html2text
pandas
lxml